+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## `0.4.0` - 2019-04-16
 ### Added
+- Smart Lock Password for the `Facebook authorization`
 
 ## `0.3.0` - 2019-04-16
 ### Added